requirements.txt
flask
flask-cors
flask-socketio
pandas
yfinance
multitasking==0.0.11
requests
websocket-client
# pandas-ta (replaced by ta_mock)
numpy
python-dotenv
schedule
pymssql
pyodbc
